+++ src/Exceptions/NotFoundException.php
@@ -2,6 +2,7 @@
 
 namespace CloudCreativity\LaravelJsonApi\Exceptions;
 
+use CloudCreativity\LaravelJsonApi\Document\Error;
 use Exception;
 use Illuminate\Http\Response;
 use Neomerx\JsonApi\Exceptions\JsonApiException;
@@ -18,5 +19,10 @@ class NotFoundException extends JsonApiException
     public function __construct($errors = [], Exception $previous = null)
     {
         parent::__construct($errors, Response::HTTP_NOT_FOUND, $previous);
+
+        $this->addError(Error::create([
+            Error::TITLE => 'Not Found',
+            Error::STATUS => Response::HTTP_NOT_FOUND,
+        ]));
     }
 }